Add handling for [] in query & fragment.
Whether [ and ] are allowed in these areas is a point of contention.
RFC3986 states in section 2.2:
``
If a reserved character is found in a URI component and no delimiting
role is known for that character, then it must be interpreted as
representing the data octet corresponding to that character's encoding
in US-ASCII.
''

But then it doesn't list these gen-delims in query or fragment, nor does
RFC3987 list them in iquery or ifragment; while they do list other
gen-delims. But, since it seems to be the practice to pass arrays in
GET-requests, and no delimiting use is known for them in these segments,
handle them as though they're allowed.